Initial Commit
Created schema.sql file to enforce the DB schema. Uses a fact table for Launches with a one to many child table for the cores to reduce clutter in the fact table. Incorporates two dimension tables holding values for the rockets and payloads.

Created a load file that uses a public GitHub API to extract the SpaceX Launches, Payloads, and Rocket data. It then transforms and loads that JSON data into the database enforcing the previously defined schema.